Remove static listening port for peer config
If the client/peer hosts a WireGuard server on 51820 they will not be able to connect to another WireGuard server themselves, as the listening port of 51820 is already used.
Removing this static port from the peer config will set the listening port to default behaviour: Automatic (port assignment). Peers/clients *should* automatically set a listening port for communication with the server, so port conflicts are avoided.