non-alphanumeric peer names are not handled gracefully
PR #134 (v1.0.20210914-ls48) changed how peer names are handled.

Prior to that release, an environment variable like the following worked:

```
- PEERS=jill-macbook,jack-chromebook,alex-nokia-g10
```

Since that release, what was then in `root/etc/cont-init.d/40-confs` and
is now in
[`root/etc/s6-overlay/s6-rc.d/init-wireguard-confs/run`](https://github.com/linuxserver/docker-wireguard/blob/65e9b6148cc72a8094e9fb8de391f0f82b69b41f/root/etc/s6-overlay/s6-rc.d/init-wireguard-confs/run#L81)
has contained this line:

```
SERVER_ALLOWEDIPS=SERVER_ALLOWEDIPS_PEER_${i}
```

When a peer name contains a character that is not allowed in a bash
variable name, that statement fails with:

```
/etc/s6-overlay/s6-rc.d/init-wireguard-confs/run: line xx: SERVER_ALLOWEDIPS_PEER_a-1: invalid variable name
```

The `run` script aborts at that point.

This problem was reported in Issue #167. The person making the report
was directed to the "(alphanumeric only)" in the
[documentation](https://github.com/linuxserver/docker-wireguard#parameters)
and closed the issue.

Further analysis suggests there are two unresolved problems:

1. The first part of the for-loop that iterates over peer names is:

	```
	  for i in ${PEERS_ARRAY[@]}; do
	    if [[ "${i}" =~ ^[0-9]+$ ]]; then
	      PEER_ID="peer${i}"
	    else
	      PEER_ID="peer_${i//[^[:alnum:]_-]/}"
	    fi
	    mkdir -p /config/${PEER_ID}
	```

	The statement after the `else` permits alphanumerics, underscores
	and hyphens, so it is not consistent with the "(alphanumeric only)"
	documentation. It would be better if the code reflected the
	documentation and was expressed as:

	```
	      PEER_ID="peer_${i//[^[:alnum:]]/}"
	```

	That change has the effect of "sanitising" peer names.
	In other words:

	```
	- PEERS=jill-macbook,jack-chromebook,alex-nokia-g10
	```

	is interpreted as if it had been written as:

	```
	- PEERS=jillmacbook,jackchromebook,alexnokiag10
	```

2. In all but four places in the body of the for-loop, `PEER_ID` is used
to refer to the "sanitised" version of the name. The other four places
refer to `${i}` which is the raw version of the name. I was not able to
construct a test case which produced inconsistencies so I don't *think*
there are problems in the wild. In any event, `PEER_ID` needs to be
used if the "sanitised" name is to make it through to peer creation
and prevent the script from aborting.

Together, both proposed changes mean that the revised:

```
SERVER_ALLOWEDIPS=SERVER_ALLOWEDIPS_PEER_${PEER_ID}
```

will always succeed (at the price of sanitising any invalid names) and
will not cause the `run` script to abort.

Signed-off-by: Phill Kelley <34226495+Paraphraser@users.noreply.github.com>